refactor: social을 socialBoard와 socialInteraction으로 분리
