+++ src/test/java/im/toduck/domain/social/domain/usecase/SocialBoardUseCaseTest.java
@@ -31,16 +31,12 @@ import im.toduck.domain.social.persistence.entity.SocialCategory;
 import im.toduck.domain.social.persistence.entity.SocialCategoryLink;
 import im.toduck.domain.social.persistence.entity.SocialImageFile;
 import im.toduck.domain.social.persistence.repository.CommentRepository;
-import im.toduck.domain.social.persistence.repository.LikeRepository;
 import im.toduck.domain.social.persistence.repository.SocialCategoryLinkRepository;
 import im.toduck.domain.social.persistence.repository.SocialImageFileRepository;
 import im.toduck.domain.social.persistence.repository.SocialRepository;
-import im.toduck.domain.social.presentation.dto.request.CommentCreateRequest;
 import im.toduck.domain.social.presentation.dto.request.SocialCreateRequest;
 import im.toduck.domain.social.presentation.dto.request.SocialUpdateRequest;
-import im.toduck.domain.social.presentation.dto.response.CommentCreateResponse;
 import im.toduck.domain.social.presentation.dto.response.CommentDto;
-import im.toduck.domain.social.presentation.dto.response.LikeCreateResponse;
 import im.toduck.domain.social.presentation.dto.response.SocialCreateResponse;
 import im.toduck.domain.social.presentation.dto.response.SocialDetailResponse;
 import im.toduck.domain.social.presentation.dto.response.SocialImageDto;
@@ -50,14 +46,13 @@ import im.toduck.fixtures.social.SocialImageFileFixtures;
 import im.toduck.fixtures.user.UserFixtures;
 import im.toduck.global.exception.CommonException;
 import im.toduck.global.exception.ExceptionCode;
-import im.toduck.global.exception.VoException;
 import im.toduck.global.presentation.dto.response.CursorPaginationResponse;
 
-public class SocialUseCaseTest extends ServiceTest {
+public class SocialBoardUseCaseTest extends ServiceTest {
 	private User USER;
 
 	@Autowired
-	private SocialUseCase socialUseCase;
+	private SocialBoardUseCase socialBoardUseCase;
 
 	@Autowired
 	private SocialRepository socialRepository;
@@ -70,9 +65,6 @@ public class SocialUseCaseTest extends ServiceTest {
 
 	@Autowired
 	private CommentRepository commentRepository;
-
-	@Autowired
-	private LikeRepository likeRepository;
 
 	@BeforeEach
 	public void setUp() {
@@ -100,7 +92,7 @@ public class SocialUseCaseTest extends ServiceTest {
 		@Test
 		void 주어진_요청에_따라_게시글을_생성할_수_있다() {
 			// when
-			SocialCreateResponse response = socialUseCase.createSocialBoard(USER.getId(), request);
+			SocialCreateResponse response = socialBoardUseCase.createSocialBoard(USER.getId(), request);
 
 			// then
 			assertSoftly(softly -> {
@@ -116,7 +108,7 @@ public class SocialUseCaseTest extends ServiceTest {
 			Long nonExistentUserId = -1L;
 
 			// when & then
-			assertThatThrownBy(() -> socialUseCase.createSocialBoard(nonExistentUserId, request))
+			assertThatThrownBy(() -> socialBoardUseCase.createSocialBoard(nonExistentUserId, request))
 				.isInstanceOf(CommonException.class)
 				.hasMessage(ExceptionCode.NOT_FOUND_USER.getMessage());
 		}
@@ -130,7 +122,7 @@ public class SocialUseCaseTest extends ServiceTest {
 			);
 
 			// when & then
-			assertThatThrownBy(() -> socialUseCase.createSocialBoard(USER.getId(), invalidRequest))
+			assertThatThrownBy(() -> socialBoardUseCase.createSocialBoard(USER.getId(), invalidRequest))
 				.isInstanceOf(CommonException.class)
 				.hasMessage(ExceptionCode.NOT_FOUND_SOCIAL_CATEGORY.getMessage());
 		}
@@ -138,80 +130,9 @@ public class SocialUseCaseTest extends ServiceTest {
 	}
 
 	@Nested
-	@DisplayName("댓글 작성시")
-	class CreateComment {
-		Social SOCIAL_BOARD;
-		String commentContent = "This is a test comment.";
+	@DisplayName("게시글 삭제시")
+	class DeleteSocialBoard {
 
-		CommentCreateRequest request = new CommentCreateRequest(commentContent);
-
-		@BeforeEach
-		public void setUp() {
-			SOCIAL_BOARD = testFixtureBuilder.buildSocial(SINGLE_SOCIAL(USER, false));
-		}
-
-		@Test
-		void 주어진_요청에_따라_댓글을_생성할_수_있다() {
-			// when
-			CommentCreateResponse response = socialUseCase.createComment(USER.getId(), SOCIAL_BOARD.getId(), request);
-
-			// then
-			assertSoftly(softly -> {
-				softly.assertThat(response).isNotNull();
-				softly.assertThat(response.commentId()).isNotNull();
-			});
-		}
-
-		@Test
-		void 사용자를_조회할_수_없는_경우_댓글_생성에_실패한다() {
-			// given
-			Long nonExistentUserId = -1L;
-
-			// when & then
-			assertThatThrownBy(() -> socialUseCase.createComment(nonExistentUserId, SOCIAL_BOARD.getId(), request))
-				.isInstanceOf(CommonException.class)
-				.hasMessage(ExceptionCode.NOT_FOUND_USER.getMessage());
-		}
-
-		@Test
-		void 소셜_게시글이_존재하지_않는_경우_댓글_생성에_실패한다() {
-			// given
-			Long nonExistentSocialBoardId = -1L;
-
-			// when & then
-			assertThatThrownBy(() -> socialUseCase.createComment(USER.getId(), nonExistentSocialBoardId, request))
-				.isInstanceOf(CommonException.class)
-				.hasMessage(ExceptionCode.NOT_FOUND_SOCIAL_BOARD.getMessage());
-		}
-
-		@Test
-		void 댓글_내용이_빈_경우_댓글_생성에_실패한다() {
-			// given
-			String emptyContent = "";
-			CommentCreateRequest emptyRequest = new CommentCreateRequest(emptyContent);
-
-			// when & then
-			assertThatThrownBy(() -> socialUseCase.createComment(USER.getId(), SOCIAL_BOARD.getId(), emptyRequest))
-				.isInstanceOf(VoException.class)
-				.hasMessage("댓글 내용은 비어 있을 수 없습니다.");
-		}
-
-		@Test
-		void 댓글_내용이_공백인_경우_댓글_생성에_실패한다() {
-			// given
-			String whitespaceContent = "   ";
-			CommentCreateRequest whitespaceRequest = new CommentCreateRequest(whitespaceContent);
-
-			// when & then
-			assertThatThrownBy(() -> socialUseCase.createComment(USER.getId(), SOCIAL_BOARD.getId(), whitespaceRequest))
-				.isInstanceOf(VoException.class)
-				.hasMessage("댓글 내용은 비어 있을 수 없습니다.");
-		}
-	}
-
-	@Nested
-	@DisplayName("게시글 댓글 삭제시")
-	class DeleteComment {
 		Social SOCIAL_BOARD;
 		Comment COMMENT;
 
@@ -222,109 +143,9 @@ public class SocialUseCaseTest extends ServiceTest {
 		}
 
 		@Test
-		@Transactional
-		void 게시글이_삭제되면_댓글은_soft_delete_된다() {
-			// when
-			socialUseCase.deleteSocialBoard(USER.getId(), SOCIAL_BOARD.getId());
-
-			// then
-			Optional<Comment> softDeletedComment = commentRepository.findById(COMMENT.getId());
-			assertSoftly(softly -> {
-				softly.assertThat(softDeletedComment).isPresent();
-				softDeletedComment.ifPresent(value -> softly.assertThat(value.getDeletedAt()).isNotNull());
-
-			});
-		}
-
-		@Test
-		void 댓글이_직접_삭제되면_hard_delete_된다() {
-			// when
-			socialUseCase.deleteComment(USER.getId(), SOCIAL_BOARD.getId(), COMMENT.getId());
-
-			// then
-			assertThat(commentRepository.findById(COMMENT.getId())).isNotPresent();
-		}
-
-		@Test
-		void 사용자를_조회할_수_없는_경우_댓글_삭제에_실패한다() {
-			// given
-			Long nonExistentUserId = -1L;
-
-			// when & then
-			assertThatThrownBy(
-				() -> socialUseCase.deleteComment(nonExistentUserId, SOCIAL_BOARD.getId(), COMMENT.getId()))
-				.isInstanceOf(CommonException.class)
-				.hasMessage(ExceptionCode.NOT_FOUND_USER.getMessage());
-		}
-
-		@Test
-		void 게시글이_존재하지_않는_경우_댓글_삭제에_실패한다() {
-			// given
-			Long nonExistentSocialBoardId = -1L;
-
-			// when & then
-			assertThatThrownBy(
-				() -> socialUseCase.deleteComment(USER.getId(), nonExistentSocialBoardId, COMMENT.getId()))
-				.isInstanceOf(CommonException.class)
-				.hasMessage(ExceptionCode.NOT_FOUND_SOCIAL_BOARD.getMessage());
-		}
-
-		@Test
-		void 댓글이_존재하지_않는_경우_댓글_삭제에_실패한다() {
-			// given
-			Long nonExistentCommentId = -1L;
-
-			// when & then
-			assertThatThrownBy(
-				() -> socialUseCase.deleteComment(USER.getId(), SOCIAL_BOARD.getId(), nonExistentCommentId))
-				.isInstanceOf(CommonException.class)
-				.hasMessage(ExceptionCode.NOT_FOUND_COMMENT.getMessage());
-		}
-
-		@Test
-		void 댓글의_소유자가_아닌_경우_삭제에_실패한다() {
-			// given
-			User ANOTHER_USER = testFixtureBuilder.buildUser(GENERAL_USER());
-			Comment ANOTHER_USER_COMMENT = testFixtureBuilder.buildComment(
-				SINGLE_COMMENT(ANOTHER_USER, SOCIAL_BOARD)
-			);
-
-			// when & then
-			assertThatThrownBy(
-				() -> socialUseCase.deleteComment(USER.getId(), SOCIAL_BOARD.getId(), ANOTHER_USER_COMMENT.getId()))
-				.isInstanceOf(CommonException.class)
-				.hasMessage(ExceptionCode.UNAUTHORIZED_ACCESS_COMMENT.getMessage());
-		}
-
-		@Test
-		void 댓글이_게시글에_속하지_않는_경우_삭제에_실패한다() {
-			// given
-			Social ANOTHER_BOARD = testFixtureBuilder.buildSocial(SINGLE_SOCIAL(USER, false));
-			Comment ANOTHER_BOARD_COMMENT = testFixtureBuilder.buildComment(SINGLE_COMMENT(USER, ANOTHER_BOARD));
-
-			// when & then
-			assertThatThrownBy(
-				() -> socialUseCase.deleteComment(USER.getId(), SOCIAL_BOARD.getId(), ANOTHER_BOARD_COMMENT.getId()))
-				.isInstanceOf(CommonException.class)
-				.hasMessage(ExceptionCode.INVALID_COMMENT_FOR_BOARD.getMessage());
-		}
-	}
-
-	@Nested
-	@DisplayName("게시글 삭제시")
-	class DeleteSocialBoard {
-
-		Social SOCIAL_BOARD;
-
-		@BeforeEach
-		void setUp() {
-			SOCIAL_BOARD = testFixtureBuilder.buildSocial(SINGLE_SOCIAL(USER, false));
-		}
-
-		@Test
 		void 주어진_요청에_따라_게시글을_삭제할_수_있다() {
 			// when
-			socialUseCase.deleteSocialBoard(USER.getId(), SOCIAL_BOARD.getId());
+			socialBoardUseCase.deleteSocialBoard(USER.getId(), SOCIAL_BOARD.getId());
 
 			// then
 			assertThat(socialRepository.findById(SOCIAL_BOARD.getId())).isEmpty();
@@ -336,7 +157,7 @@ public class SocialUseCaseTest extends ServiceTest {
 			Long nonExistentUserId = -1L;
 
 			// when & then
-			assertThatThrownBy(() -> socialUseCase.deleteSocialBoard(nonExistentUserId, SOCIAL_BOARD.getId()))
+			assertThatThrownBy(() -> socialBoardUseCase.deleteSocialBoard(nonExistentUserId, SOCIAL_BOARD.getId()))
 				.isInstanceOf(CommonException.class)
 				.hasMessage(ExceptionCode.NOT_FOUND_USER.getMessage());
 		}
@@ -347,7 +168,7 @@ public class SocialUseCaseTest extends ServiceTest {
 			Long nonExistentSocialBoardId = -1L;
 
 			// when & then
-			assertThatThrownBy(() -> socialUseCase.deleteSocialBoard(USER.getId(), nonExistentSocialBoardId))
+			assertThatThrownBy(() -> socialBoardUseCase.deleteSocialBoard(USER.getId(), nonExistentSocialBoardId))
 				.isInstanceOf(CommonException.class)
 				.hasMessage(ExceptionCode.NOT_FOUND_SOCIAL_BOARD.getMessage());
 		}
@@ -359,9 +180,24 @@ public class SocialUseCaseTest extends ServiceTest {
 			Long socialBoardId = SOCIAL_BOARD.getId();
 
 			// when & then
-			assertThatThrownBy(() -> socialUseCase.deleteSocialBoard(ANOTHER_USER.getId(), socialBoardId))
+			assertThatThrownBy(() -> socialBoardUseCase.deleteSocialBoard(ANOTHER_USER.getId(), socialBoardId))
 				.isInstanceOf(CommonException.class)
 				.hasMessage(ExceptionCode.UNAUTHORIZED_ACCESS_SOCIAL_BOARD.getMessage());
+		}
+
+		@Test
+		@Transactional
+		void 게시글이_삭제되면_댓글은_soft_delete_된다() {
+			// when
+			socialBoardUseCase.deleteSocialBoard(USER.getId(), SOCIAL_BOARD.getId());
+
+			// then
+			Optional<Comment> softDeletedComment = commentRepository.findById(COMMENT.getId());
+			assertSoftly(softly -> {
+				softly.assertThat(softDeletedComment).isPresent();
+				softDeletedComment.ifPresent(value -> softly.assertThat(value.getDeletedAt()).isNotNull());
+
+			});
 		}
 	}
 
@@ -426,7 +262,7 @@ public class SocialUseCaseTest extends ServiceTest {
 			);
 
 			// when
-			socialUseCase.updateSocialBoard(USER.getId(), SOCIAL_BOARD.getId(), updateRequest);
+			socialBoardUseCase.updateSocialBoard(USER.getId(), SOCIAL_BOARD.getId(), updateRequest);
 
 			// then
 			Social updatedSocialBoard = socialRepository.findById(SOCIAL_BOARD.getId()).orElseThrow();
@@ -486,7 +322,7 @@ public class SocialUseCaseTest extends ServiceTest {
 
 			// when & then
 			assertThatThrownBy(
-				() -> socialUseCase.updateSocialBoard(USER.getId(), nonExistentSocialBoardId, updateRequest))
+				() -> socialBoardUseCase.updateSocialBoard(USER.getId(), nonExistentSocialBoardId, updateRequest))
 				.isInstanceOf(CommonException.class)
 				.hasMessage(ExceptionCode.NOT_FOUND_SOCIAL_BOARD.getMessage());
 		}
@@ -504,7 +340,7 @@ public class SocialUseCaseTest extends ServiceTest {
 
 			// when & then
 			assertThatThrownBy(
-				() -> socialUseCase.updateSocialBoard(ANOTHER_USER.getId(), SOCIAL_BOARD.getId(), updateRequest))
+				() -> socialBoardUseCase.updateSocialBoard(ANOTHER_USER.getId(), SOCIAL_BOARD.getId(), updateRequest))
 				.isInstanceOf(CommonException.class)
 				.hasMessage(ExceptionCode.UNAUTHORIZED_ACCESS_SOCIAL_BOARD.getMessage());
 		}
@@ -520,7 +356,8 @@ public class SocialUseCaseTest extends ServiceTest {
 			);
 
 			// when & then
-			assertThatThrownBy(() -> socialUseCase.updateSocialBoard(USER.getId(), SOCIAL_BOARD.getId(), updateRequest))
+			assertThatThrownBy(
+				() -> socialBoardUseCase.updateSocialBoard(USER.getId(), SOCIAL_BOARD.getId(), updateRequest))
 				.isInstanceOf(CommonException.class)
 				.hasMessage(ExceptionCode.NOT_FOUND_SOCIAL_CATEGORY.getMessage());
 		}
@@ -536,107 +373,10 @@ public class SocialUseCaseTest extends ServiceTest {
 			);
 
 			// when & then
-			assertThatThrownBy(() -> socialUseCase.updateSocialBoard(USER.getId(), SOCIAL_BOARD.getId(), updateRequest))
+			assertThatThrownBy(
+				() -> socialBoardUseCase.updateSocialBoard(USER.getId(), SOCIAL_BOARD.getId(), updateRequest))
 				.isInstanceOf(CommonException.class)
 				.hasMessage(ExceptionCode.EMPTY_SOCIAL_CATEGORY_LIST.getMessage());
-		}
-	}
-
-	@Nested
-	@DisplayName("게시글 좋아요 생성시")
-	class CreateLikeTest {
-		Social SOCIAL_BOARD;
-
-		@BeforeEach
-		void setUp() {
-			SOCIAL_BOARD = testFixtureBuilder.buildSocial(SINGLE_SOCIAL(USER, false));
-		}
-
-		@Test
-		void 게시글에_좋아요를_성공적으로_생성한다() {
-			// given
-			int beforeLikeCount = SOCIAL_BOARD.getLikeCount();
-
-			// when
-			LikeCreateResponse response = socialUseCase.createLike(USER.getId(), SOCIAL_BOARD.getId());
-
-			// then
-			Like createdLike = likeRepository.findByUserAndSocial(USER, SOCIAL_BOARD).orElseThrow();
-			Optional<Social> afterSocialBoard = socialRepository.findById(SOCIAL_BOARD.getId());
-			assertSoftly(softly -> {
-				assertThat(createdLike).isNotNull();
-				assertThat(response.likeId()).isNotNull();
-				afterSocialBoard.ifPresent(
-					social -> softly.assertThat(social.getLikeCount()).isEqualTo(beforeLikeCount + 1)
-				);
-
-			});
-		}
-
-		@Test
-		void 이미_좋아요를_누른_경우_예외를_발생시킨다() {
-			// given
-			LikeCreateResponse response = socialUseCase.createLike(USER.getId(), SOCIAL_BOARD.getId());
-
-			// when & then
-			assertThatThrownBy(() -> socialUseCase.createLike(USER.getId(), SOCIAL_BOARD.getId()))
-				.isInstanceOf(CommonException.class)
-				.hasMessage(ExceptionCode.EXISTS_LIKE.getMessage());
-		}
-
-		@Test
-		void 존재하지_않는_게시글에_좋아요를_누를_경우_예외를_발생시킨다() {
-			// given
-			Long nonExistentSocialId = -1L;
-
-			// when & then
-			assertThatThrownBy(() -> socialUseCase.createLike(USER.getId(), nonExistentSocialId))
-				.isInstanceOf(CommonException.class)
-				.hasMessage(ExceptionCode.NOT_FOUND_SOCIAL_BOARD.getMessage());
-		}
-	}
-
-	@Nested
-	@DisplayName("게시글 좋아요 취소시")
-	class DeleteLikeTest {
-		Social SOCIAL_BOARD;
-
-		@BeforeEach
-		void setUp() {
-			SOCIAL_BOARD = testFixtureBuilder.buildSocial(SINGLE_SOCIAL(USER, false));
-		}
-
-		@Test
-		void 게시글에_좋아요를_성공적으로_취소한다() {
-			// given
-			Like LIKE = testFixtureBuilder.buildLike(LIKE(USER, SOCIAL_BOARD));
-
-			// when
-			socialUseCase.deleteLike(USER.getId(), SOCIAL_BOARD.getId());
-
-			// then
-			assertThat(likeRepository.findByUserAndSocial(USER, SOCIAL_BOARD)).isNotPresent();
-			assertThat(SOCIAL_BOARD.getLikeCount()).isEqualTo(0);
-		}
-
-		@Test
-		void 좋아요가_존재하지_않는_경우_예외를_발생시킨다() {
-
-			// when & then
-			assertThatThrownBy(() -> socialUseCase.deleteLike(USER.getId(), SOCIAL_BOARD.getId()))
-				.isInstanceOf(CommonException.class)
-				.hasMessage(ExceptionCode.NOT_FOUND_LIKE.getMessage());
-		}
-
-		@Test
-		void 존재하지_않는_게시글에_좋아요_취소를_시도하면_예외를_발생시킨다() {
-			// given
-			Long nonExistentSocialId = -1L;
-
-			// when & then
-			assertThatThrownBy(() -> socialUseCase.deleteLike(USER.getId(), nonExistentSocialId))
-				.isInstanceOf(CommonException.class)
-				.hasMessage(ExceptionCode.NOT_FOUND_SOCIAL_BOARD.getMessage());
 		}
 	}
 
@@ -670,7 +410,7 @@ public class SocialUseCaseTest extends ServiceTest {
 		@Test
 		void 게시글_단건_조회에_성공한다() {
 			// when
-			SocialDetailResponse response = socialUseCase.getSocialDetail(USER.getId(), SOCIAL_BOARD.getId());
+			SocialDetailResponse response = socialBoardUseCase.getSocialDetail(USER.getId(), SOCIAL_BOARD.getId());
 
 			// then
 			assertSoftly(softly -> {
@@ -702,7 +442,7 @@ public class SocialUseCaseTest extends ServiceTest {
 			Long nonExistentSocialId = -1L;
 
 			// when & then
-			assertThatThrownBy(() -> socialUseCase.getSocialDetail(USER.getId(), nonExistentSocialId))
+			assertThatThrownBy(() -> socialBoardUseCase.getSocialDetail(USER.getId(), nonExistentSocialId))
 				.isInstanceOf(CommonException.class)
 				.hasMessage(ExceptionCode.NOT_FOUND_SOCIAL_BOARD.getMessage());
 		}
@@ -713,7 +453,7 @@ public class SocialUseCaseTest extends ServiceTest {
 			Long nonExistentUserId = -1L;
 
 			// when & then
-			assertThatThrownBy(() -> socialUseCase.getSocialDetail(nonExistentUserId, SOCIAL_BOARD.getId()))
+			assertThatThrownBy(() -> socialBoardUseCase.getSocialDetail(nonExistentUserId, SOCIAL_BOARD.getId()))
 				.isInstanceOf(CommonException.class)
 				.hasMessage(ExceptionCode.NOT_FOUND_USER.getMessage());
 		}
@@ -721,7 +461,7 @@ public class SocialUseCaseTest extends ServiceTest {
 		@Test
 		void 차단된_사용자의_게시글을_조회하려_할_경우_예외를_발생시킨다() {
 			// when & then
-			assertThatThrownBy(() -> socialUseCase.getSocialDetail(USER.getId(), BLOCKED_USER_SOCIAL.getId()))
+			assertThatThrownBy(() -> socialBoardUseCase.getSocialDetail(USER.getId(), BLOCKED_USER_SOCIAL.getId()))
 				.isInstanceOf(CommonException.class)
 				.hasMessage(ExceptionCode.BLOCKED_USER_SOCIAL_ACCESS.getMessage());
 		}
@@ -740,7 +480,8 @@ public class SocialUseCaseTest extends ServiceTest {
 			Integer limit = 10;
 
 			// when
-			CursorPaginationResponse<SocialResponse> response = socialUseCase.getSocials(USER.getId(), cursor, limit);
+			CursorPaginationResponse<SocialResponse> response = socialBoardUseCase.getSocials(USER.getId(), cursor,
+				limit);
 
 			// then
 			assertSoftly(softly -> {
@@ -759,10 +500,9 @@ public class SocialUseCaseTest extends ServiceTest {
 			Integer limit = 10;
 
 			// when & then
-			assertThatThrownBy(() -> socialUseCase.getSocials(nonExistentUserId, cursor, limit))
+			assertThatThrownBy(() -> socialBoardUseCase.getSocials(nonExistentUserId, cursor, limit))
 				.isInstanceOf(CommonException.class)
 				.hasMessage(ExceptionCode.NOT_FOUND_USER.getMessage());
 		}
 	}
-
 }
